+++ perllib/FixMyStreet/SendReport/Email.pm
@@ -71,7 +71,7 @@ sub send {
     }
 
     my ($verbose, $nomail) = CronFns::options();
-    my $cobrand = FixMyStreet::Cobrand->get_class_for_moniker($row->cobrand)->new();
+    my $cobrand = $row->get_cobrand_logged;
     $cobrand = $cobrand->call_hook(get_body_handler_for_problem => $row) || $cobrand;
 
     my $params = {